refactor(storybook): modularize theme handling and enhance documentation support
- Refactored ThemeDecorator and added CustomDocsContainer to manage dark mode synchronization in Storybook.
- Updated preview.tsx to utilize the new CustomDocsContainer for documentation theming.
- Changed APCACheck export to named export for consistency.
- Removed redundant theme handling logic from ThemeDecorator, improving code clarity.
