+++ packages/nimbus/.storybook/preview.tsx
@@ -1,10 +1,9 @@
 import React, { useEffect, useState } from "react";
-import type { Preview, StoryContext } from "@storybook/react-vite";
-import { NimbusProvider } from "../src";
-import { DARK_MODE_EVENT_NAME } from "@vueless/storybook-dark-mode";
-import { addons } from "storybook/preview-api";
+import type { Preview } from "@storybook/react-vite";
 
-import APCACheck from "./apca-check";
+import { APCACheck } from "./apca-check";
+import { CustomDocsContainer } from "./docs-container";
+import { ThemeDecorator } from "./theme-decorator";
 
 const apca = APCACheck("custom", (fontSize: string) => {
   const size = parseFloat(fontSize);
@@ -15,47 +14,6 @@ const apca = APCACheck("custom", (fontSize: string) => {
       return 60;
   }
 });
-
-// get channel to listen to event emitter
-const channel = addons.getChannel();
-
-const ThemeDecorator = ({
-  children,
-  context,
-}: {
-  children: React.ReactNode;
-  context: StoryContext;
-}) => {
-  const [isDark, setDark] = useState(false);
-  const theme = isDark ? "dark" : "light";
-  const { locale } = context.globals;
-
-  useEffect(() => {
-    const { current } = JSON.parse(
-      // TODO: find out if there is a more elegant solution
-      localStorage.getItem("sb-addon-themes-3") || "{}"
-    );
-
-    setDark(current === "dark");
-
-    channel.on(DARK_MODE_EVENT_NAME, (darkMode) => {
-      setDark(darkMode);
-    });
-
-    return () => channel.off(DARK_MODE_EVENT_NAME, setDark);
-  }, [channel]);
-
-  useEffect(() => {
-    document.documentElement.classList.remove("light", "dark");
-    document.documentElement.classList.add(theme);
-  }, [theme]);
-
-  return (
-    <NimbusProvider locale={locale} defaultTheme={theme}>
-      {children}
-    </NimbusProvider>
-  );
-};
 
 const preview: Preview = {
   parameters: {
@@ -88,6 +46,9 @@ const preview: Preview = {
       options: {},
       // fail the test runner if a11y violations are found
       test: "error",
+    },
+    docs: {
+      container: CustomDocsContainer,
     },
   },
   tags: ["autodocs", "a11y-test"],
